Move date formating utils to its own file
