+++ application/ui/src/features/models/model-listing/components/model-row/model-row.component.tsx
@@ -8,11 +8,11 @@ import type {
     Model,
     ModelArchitectureWithPerformanceCategory,
 } from '../../../../../constants/shared-types';
+import { formatTrainingDateTime } from '../../../../../shared/date-utils';
 import { formatBytes } from '../../../../../shared/util';
 import { GRID_COLUMNS } from '../../constants';
 import { AccuracyIndicator } from '../../model-variants/accuracy-indicator/accuracy-indicator.component';
 import { type GroupByMode } from '../../types';
-import { formatTrainingDateTime } from '../../utils/date-formatting';
 import { hasDeletedWeights, isFailedModel } from '../../utils/utils';
 import { ParentRevisionModel } from '../parent-revision-model.component';
 import { ArchitectureColumn } from './architecture-column.component';
